airflow-operator/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-weaver-airflow"
version = "0.2.0"
description = "Airflow operator for Data Weaver pipelines"
requires-python = ">=3.9"
dependencies = [
    "apache-airflow>=2.7.0",
]

[project.urls]
Repository = "https://github.com/netsirius/data-weaver"
